refactoring the openarc and openai endpoints to their own py files
